Organize group-related components
